[KodiXml] Simplify handling of different Kodi versions
