+++ src/media_centers/KodiXml.h
@@ -102,11 +102,6 @@ public:
         StreamDetails* streamDetails,
         const QVector<Subtitle*>& subtitles,
         bool hasStreamDetails);
-    static void setListValue(QDomDocument& doc, const QString& name, const QStringList& values);
-    static QDomElement addTextValue(QDomDocument& doc, const QString& name, const QString& value);
-    static void appendXmlNode(QDomDocument& doc, QDomNode& node);
-    static void removeChildNodes(QDomDocument& doc, const QString& name);
-    static QDomElement setTextValue(QDomDocument& doc, const QString& name, const QString& value);
 
     static void writeStringsAsOneTagEach(QXmlStreamWriter& xml, const QString& name, const QStringList& list);
