+++ src/media_centers/kodi/ConcertXmlWriter.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "media_centers/kodi/KodiXmlWriter.h"
+
 #include <QByteArray>
 
 class Concert;
@@ -7,11 +9,22 @@ class Concert;
 namespace mediaelch {
 namespace kodi {
 
-class ConcertXmlWriter
+class ConcertXmlWriter : public KodiXmlWriter
 {
 public:
+    ConcertXmlWriter(KodiVersion version) : KodiXmlWriter(std::move(version)) {}
     virtual ~ConcertXmlWriter() = default;
     virtual QByteArray getConcertXml(bool testMode = false) = 0;
+};
+
+class ConcertXmlWriterGeneric : public ConcertXmlWriter
+{
+public:
+    ConcertXmlWriterGeneric(KodiVersion version, Concert& concert);
+    QByteArray getConcertXml(bool testMode = false) override;
+
+private:
+    Concert& m_concert;
 };
 
 } // namespace kodi
