+++ src/media_centers/kodi/EpisodeXmlWriter.h
@@ -1,21 +1,38 @@
 #pragma once
 
-#include <QByteArray>
+#include "media_centers/kodi/KodiXmlWriter.h"
 
-class Episode;
+#include <QByteArray>
+#include <QXmlStreamWriter>
+
+class TvShowEpisode;
 
 namespace mediaelch {
 namespace kodi {
 
-class EpisodeXmlWriter
+class EpisodeXmlWriter : public KodiXmlWriter
 {
 public:
+    EpisodeXmlWriter(KodiVersion version) : KodiXmlWriter(std::move(version)) {}
     virtual ~EpisodeXmlWriter() = default;
     virtual QByteArray getEpisodeXml(bool testMode = false) = 0;
     /// Get the episode's XML content with an extra root element: <episodes>
     /// Used for testing to allow multi-episodes in one DOM.
     virtual QString getEpisodeXmlWithSingleRoot(bool testMode);
 };
+
+class EpisodeXmlWriterGeneric : public EpisodeXmlWriter
+{
+public:
+    EpisodeXmlWriterGeneric(KodiVersion version, const QVector<TvShowEpisode*>& episodes);
+    QByteArray getEpisodeXml(bool testMode = false) override;
+
+private:
+    void writeSingleEpisodeDetails(QXmlStreamWriter& xml, TvShowEpisode* episode, bool testMode);
+
+    const QVector<TvShowEpisode*> m_episodes;
+};
+
 
 } // namespace kodi
 } // namespace mediaelch
