+++ src/media_centers/kodi/MovieXmlWriter.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "media_centers/kodi/KodiXmlWriter.h"
+
 #include <QByteArray>
 
 class Movie;
@@ -7,11 +9,22 @@ class Movie;
 namespace mediaelch {
 namespace kodi {
 
-class MovieXmlWriter
+class MovieXmlWriter : public KodiXmlWriter
 {
 public:
+    MovieXmlWriter(KodiVersion version) : KodiXmlWriter(std::move(version)) {}
     virtual ~MovieXmlWriter() = default;
     virtual QByteArray getMovieXml(bool testMode = false) = 0;
+};
+
+class MovieXmlWriterGeneric : public MovieXmlWriter
+{
+public:
+    MovieXmlWriterGeneric(KodiVersion version, Movie& movie);
+    QByteArray getMovieXml(bool testMode = false) override;
+
+private:
+    Movie& m_movie;
 };
 
 } // namespace kodi
