--- src/media_centers/kodi/v17/AlbumXmlWriterV17.h
@@ -1,29 +1,0 @@
-#pragma once
-
-#include "media_centers/kodi/AlbumXmlWriter.h"
-
-#include <QByteArray>
-#include <QXmlStreamWriter>
-
-class Album;
-
-namespace mediaelch {
-namespace kodi {
-
-class AlbumXmlWriterV17 : public AlbumXmlWriter
-{
-public:
-    explicit AlbumXmlWriterV17(Album& album);
-    QByteArray getAlbumXml(bool testMode = false) override;
-
-private:
-    void writeAlbumTags(QXmlStreamWriter& xml);
-
-private:
-    Album& m_album;
-
-    void writeArtistCredits(QXmlStreamWriter& xml);
-};
-
-} // namespace kodi
-} // namespace mediaelch
