--- src/media_centers/kodi/v17/ConcertXmlWriterV17.h
@@ -1,23 +1,0 @@
-#pragma once
-
-#include "media_centers/kodi/ConcertXmlWriter.h"
-
-#include <QByteArray>
-
-class Concert;
-
-namespace mediaelch {
-namespace kodi {
-
-class ConcertXmlWriterV17 : public ConcertXmlWriter
-{
-public:
-    ConcertXmlWriterV17(Concert& concert);
-    QByteArray getConcertXml(bool testMode = false) override;
-
-private:
-    Concert& m_concert;
-};
-
-} // namespace kodi
-} // namespace mediaelch
