--- src/media_centers/kodi/v17/MovieXmlWriterV17.h
@@ -1,23 +1,0 @@
-#pragma once
-
-#include "media_centers/kodi/MovieXmlWriter.h"
-
-#include <QByteArray>
-
-class Movie;
-
-namespace mediaelch {
-namespace kodi {
-
-class MovieXmlWriterV17 : public MovieXmlWriter
-{
-public:
-    MovieXmlWriterV17(Movie& movie);
-    QByteArray getMovieXml(bool testMode = false) override;
-
-private:
-    Movie& m_movie;
-};
-
-} // namespace kodi
-} // namespace mediaelch
