--- src/media_centers/kodi/v18/AlbumXmlWriterV18.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include "media_centers/kodi/v17/AlbumXmlWriterV17.h"
-
-namespace mediaelch {
-namespace kodi {
-
-// currently the same
-using AlbumXmlWriterV18 = AlbumXmlWriterV17;
-
-} // namespace kodi
-} // namespace mediaelch
