--- src/media_centers/kodi/v18/ArtistXmlWriterV18.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include "media_centers/kodi/v17/ArtistXmlWriterV17.h"
-
-namespace mediaelch {
-namespace kodi {
-
-// currently the same
-using ArtistXmlWriterV18 = ArtistXmlWriterV17;
-
-} // namespace kodi
-} // namespace mediaelch
