--- src/media_centers/kodi/v18/ConcertXmlWriterV18.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include "media_centers/kodi/v17/ConcertXmlWriterV17.h"
-
-namespace mediaelch {
-namespace kodi {
-
-// currently the same
-using ConcertXmlWriterV18 = ConcertXmlWriterV17;
-
-} // namespace kodi
-} // namespace mediaelch
