--- src/media_centers/kodi/v18/EpisodeXmlWriterV18.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include "media_centers/kodi/v17/EpisodeXmlWriterV17.h"
-
-namespace mediaelch {
-namespace kodi {
-
-// currently the same
-using EpisodeXmlWriterV18 = EpisodeXmlWriterV17;
-
-} // namespace kodi
-} // namespace mediaelch
