--- src/media_centers/kodi/v18/MovieXmlWriterV18.cpp
@@ -1,9 +1,0 @@
-#include "media_centers/kodi/v18/MovieXmlWriterV18.h"
-
-namespace mediaelch {
-namespace kodi {
-
-// no-op
-
-} // namespace kodi
-} // namespace mediaelch
