--- src/media_centers/kodi/v18/MovieXmlWriterV18.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include "media_centers/kodi/v17/MovieXmlWriterV17.h"
-
-namespace mediaelch {
-namespace kodi {
-
-// currently the same
-using MovieXmlWriterV18 = MovieXmlWriterV17;
-
-} // namespace kodi
-} // namespace mediaelch
