--- src/media_centers/kodi/v18/TvShowXmlWriterV18.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include "media_centers/kodi/v17/TvShowXmlWriterV17.h"
-
-namespace mediaelch {
-namespace kodi {
-
-// currently the same
-using TvShowXmlWriterV18 = TvShowXmlWriterV17;
-
-} // namespace kodi
-} // namespace mediaelch
