+++ test/integration/media_centers/testKodi_v18_episode.cpp
@@ -1,7 +1,7 @@
 #include "test/test_helpers.h"
 
 #include "media_centers/kodi/EpisodeXmlReader.h"
-#include "media_centers/kodi/v18/EpisodeXmlWriterV18.h"
+#include "media_centers/kodi/EpisodeXmlWriter.h"
 #include "settings/Settings.h"
 #include "test/integration/resource_dir.h"
 #include "tv_shows/TvShowEpisode.h"
@@ -32,7 +32,7 @@ static void createAndCompareSingleEpisode(const QString& filename, Callback call
 
     callback(episode);
 
-    mediaelch::kodi::EpisodeXmlWriterV18 writer({&episode});
+    mediaelch::kodi::EpisodeXmlWriterGeneric writer(mediaelch::KodiVersion(18), {&episode});
     QString actual = writer.getEpisodeXmlWithSingleRoot(true).trimmed();
     writeTempFile(filename, actual);
     checkSameXml(episodeContent, actual);
@@ -61,7 +61,7 @@ static void createAndCompareMultiEpisode(const QString& filename, Callback callb
 
     callback(episodesPointer);
 
-    mediaelch::kodi::EpisodeXmlWriterV18 writer(episodesPointer);
+    mediaelch::kodi::EpisodeXmlWriterGeneric writer(mediaelch::KodiVersion(18), episodesPointer);
     QString actual = writer.getEpisodeXmlWithSingleRoot(true).trimmed();
     writeTempFile(filename, actual);
     checkSameXml(episodeContent, actual);
@@ -79,7 +79,7 @@ TEST_CASE("Episode XML writer for Kodi v18", "[data][tvshow][kodi][nfo]")
         QString filename = "show/kodi_v18_episode_empty.nfo";
         CAPTURE(filename);
 
-        mediaelch::kodi::EpisodeXmlWriterV18 writer({&episode});
+        mediaelch::kodi::EpisodeXmlWriterGeneric writer(mediaelch::KodiVersion(18), {&episode});
         QString actual = writer.getEpisodeXmlWithSingleRoot(true).trimmed();
         writeTempFile(filename, actual);
         checkSameXml(getFileContent(filename), actual);
@@ -92,7 +92,7 @@ TEST_CASE("Episode XML writer for Kodi v18", "[data][tvshow][kodi][nfo]")
         QString filename = "show/kodi_v18_episode_multi_empty.nfo";
         CAPTURE(filename);
 
-        mediaelch::kodi::EpisodeXmlWriterV18 writer({&episode1, &episode2});
+        mediaelch::kodi::EpisodeXmlWriterGeneric writer(mediaelch::KodiVersion(18), {&episode1, &episode2});
         QString actual = writer.getEpisodeXmlWithSingleRoot(true).trimmed();
         writeTempFile(filename, actual);
         checkSameXml(getFileContent(filename), actual);
@@ -113,7 +113,7 @@ TEST_CASE("Episode XML writer for Kodi v18", "[data][tvshow][kodi][nfo]")
         doc.setContent(episodeContent);
         reader.parseNfoDom(doc.elementsByTagName("episodedetails").at(0).toElement());
 
-        mediaelch::kodi::EpisodeXmlWriterV18 writer({&episode});
+        mediaelch::kodi::EpisodeXmlWriterGeneric writer(mediaelch::KodiVersion(18), {&episode});
         QString actual = writer.getEpisodeXmlWithSingleRoot(true).trimmed();
         writeTempFile(filename, actual);
         checkSameXml(episodeContent, actual);
