+++ test/integration/media_centers/testKodi_v18_movie.cpp
@@ -2,7 +2,7 @@
 
 #include "media_centers/KodiXml.h"
 #include "media_centers/kodi/MovieXmlReader.h"
-#include "media_centers/kodi/v18/MovieXmlWriterV18.h"
+#include "media_centers/kodi/MovieXmlWriter.h"
 #include "test/integration/resource_dir.h"
 
 #include <QDateTime>
@@ -29,7 +29,7 @@ static void createAndCompareMovie(const QString& filename, Callback callback)
 
     callback(movie);
 
-    mediaelch::kodi::MovieXmlWriterV18 writer(movie);
+    mediaelch::kodi::MovieXmlWriterGeneric writer(mediaelch::KodiVersion(18), movie);
     QString actual = writer.getMovieXml(true).trimmed();
     writeTempFile(filename, actual);
     checkSameXml(movieContent, actual);
@@ -43,7 +43,7 @@ TEST_CASE("Movie XML writer for Kodi v18", "[data][movie][kodi][nfo]")
         QString filename = "movie/kodi_v18_movie_empty.nfo";
         CAPTURE(filename);
 
-        mediaelch::kodi::MovieXmlWriterV18 writer(movie);
+        mediaelch::kodi::MovieXmlWriterGeneric writer(mediaelch::KodiVersion(18), movie);
         QString actual = writer.getMovieXml(true).trimmed();
         writeTempFile(filename, actual);
         checkSameXml(getFileContent(filename), actual);
@@ -191,7 +191,7 @@ TEST_CASE("Movie XML writer for Kodi v18", "[data][movie][kodi][nfo]")
         // TODO: order
         movie.addActor(actor);
 
-        mediaelch::kodi::MovieXmlWriterV18 writer(movie);
+        mediaelch::kodi::MovieXmlWriterGeneric writer(mediaelch::KodiVersion(18), movie);
 
         QString actual = writer.getMovieXml(true).trimmed();
         QString filename = "movie/kodi_v18_movie_all.nfo";
